% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bird.mig.status.R
\docType{data}
\name{Bird.mig.status}
\alias{Bird.mig.status}
\title{Information on avian species' migratory status, extracted from BirdLife International.}
\format{A \code{data.frame} object with 303 rows and two columns}
\usage{
data("Bird.mig.status")
}
\description{
A categorical variable describing the broad migratory status of each species in
\code{BBS.occurrences} was extracted from BirdLife International using web scraping. The
entire workflow is shown in the \code{Appendix S7} script.
}
\keyword{datasets}
