% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bird.traits.R
\docType{data}
\name{Bird.traits}
\alias{Bird.traits}
\title{Information on avian species' migratory status (extracted from BirdLife International),
habitat and diet diversity (extracted from EltonTraits), clutch size and rarity}
\format{A \code{data.frame} object with 303 rows and 13 columns}
\usage{
data("Bird.traits")
}
\description{
A \code{dataframe} containing variables describing the broad migratory status of each species in
\code{BBS.occurrences} (extracted from BirdLife International using web scraping) as well as
derived indices of species' habitat use and nesting behaviour (extracted from the
Land Condition Trend Analysis Avian Database), diet and habitat diversities
(calculated as a Shannon index using data in the EltonTraits database),
clutch size (extracted from the global avian life
history traits dataset) and species rarity (scaled variable representing
the proportional number of observations in \code{BBS.occurrences} in which a species
was observed). The
entire workflow is shown in the \code{Appendix S7} script.
}
\keyword{datasets}
