% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdTree.syns.R
\docType{data}
\name{BirdTree.syns}
\alias{BirdTree.syns}
\title{Synonyms for avian species, along with their matching BirdTree.org taxonomy}
\format{An object of class \code{data.frame} with 3096 rows and 3 columns.}
\usage{
data(BirdTree.syns)
}
\description{
A work-in-progress file to store possible synonyms used for
avian species, along with their verified BirdTree.org taxonomy. This
file is loaded in the \code{validateSpecies} script
}
\keyword{datasets}
