% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanStormData.R
\name{cleanStormData}
\alias{cleanStormData}
\title{Clean downloaded NOAA storm data and return the top proportion of most damaging
storms}
\usage{
cleanStormData(Stormdata.raw, damage.threshold)
}
\arguments{
\item{Stormdata.raw}{A \code{data.table} object of raw NOAA storm data, downloaded from
climage.gov}

\item{damage.threshold}{A positive numeric value stating the percentage of most damaging storms
to return. For example, if \code{damage.threshold == 0.25}, then only storms registering in
the top 25 percent of total property and crop damages will be returned. Default is to return
the top third of most damaging storms \code{damage.threshold == 0.33}}
}
\value{
A \code{data.frame} object
}
\description{
Clean downloaded NOAA storm data and return the top proportion of most damaging
storms
}
