% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRegions.R
\name{createRegions}
\alias{createRegions}
\title{Use hierarchical clustering to group sites into biogeographical regions
based on similarity in avian species composition}
\usage{
createRegions(dist_weights, n_cores)
}
\arguments{
\item{dist_weights}{Optional \code{list} giving relative weights (as integers)
for the six different distance matrices when creating the final weighted matrix. The
six matrices are: species composition, latitude, longitude, mean annual precipitation,
mean minimum temperature and mean maximum temperature. Default is to only use
species composition, latitude, longitude with equal weights}

\item{n_cores}{Positive integer stating the number of processing cores to split the job across.
Default is \code{parallel::detect_cores() - 1}}
}
\value{
A \code{dataframe} that contains the \code{\link{Site.descriptors}} along with a new column
describing the bioregions
}
\description{
Use hierarchical clustering to group sites into biogeographical regions
based on similarity in avian species composition
}
\details{
Distance matrices are created for each of the six site-level parameters. Species
composition is calculated by considering the presence-absence of species at each site, where
a presence is recorded only if that species has been observed in that site on at least five
of the seven total BBS surveys. A mean distance is then calculated from the six different matrices
using the specified weights in \code{dist_weights}. Hierachical clustering is then performed and the
resulting dendrogram is cut to group sites into broader biogeographical regions. Separate grouping
is performed for each flyway in \code{\link{Site.descriptors}}
}
