% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdleModel.R
\name{hurdleModel}
\alias{hurdleModel}
\title{Run binomial logistic models to predict species' occurrences and gaussian CRFs
to predict species' local abundances.}
\usage{
hurdleModel(flyway, n_reps, n_cores)
}
\arguments{
\item{flyway}{A \code{character} string representing the name of the specified flyway to model}

\item{n_reps}{Positive \code{integer} representing the number of times to repeat
10-fold \code{\link[glmnet]{cv.glmnet}} regularized regressions (default is \code{10})}

\item{n_cores}{Positive integer stating the number of processing cores to split the job across.
Default is \code{parallel::detect_cores() - 1}}
}
\description{
Run binomial logistic models to predict species' occurrences and gaussian CRFs
to predict species' local abundances.
}
\details{
This function runs the hurdle model by first estimating predictors of species' occurrence
probability using \code{lassoBinomial} and then fitting a conditional random field to species'
scaled abundances using \code{lassoGaussian}. Each of these models is run \code{n_reps} times
to allow for uncertainty in coefficient estimates. Predictions are then used to estimate model fits and
predict community network \emph{B}'os and species' eigencentrality in each site. All models are
run for sites within a unique bioregion, ensuring that network metrics reflect variation among
sites with similar climate and species compositions.
}
