% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanPRISM.R
\name{meanPRISM}
\alias{meanPRISM}
\title{Extract mean monthly PRISM climate values for buffers around lat/long points}
\usage{
meanPRISM(points, var_name, buffer, filepath, n_cores, delete_files)
}
\arguments{
\item{points}{A \code{dataframe} of points with columns \code{Latitude} and \code{Longitude}}

\item{var_name}{A \code{character} string specifying the name of the climate variable. If not provided,
the variable's column name in the returned \code{dataframe} will be \code{mean.value}. Note that some
particular point * month measurements may be missing, and so \code{NA}s will be returned for these. A
warnign message will be printed if this occurs.}

\item{buffer}{Positive numeric value stating the length (in meters) of the desired buffer radius.
Mean values of the climate variable will be calculated for each point from within this buffer}

\item{filepath}{A \code{character} string stating the path to the folder where unzipped
PRISM files are stored (downloaded using \code{\link[prism]{get_prism_monthlys}})}

\item{n_cores}{Positive integer stating the number of processing cores to split the job across.
Default is \code{parallel::detect_cores() - 1}}

\item{delete_files}{Logical describing whether or not to delete raw PRISM data folders once the job
is complete. Default is \code{FALSE}}
}
\description{
Extract mean monthly PRISM climate values for buffers around lat/long points
}
