% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkModel.R
\name{networkModel}
\alias{networkModel}
\title{Run linear mixed models to identify predictors of species' eigencentralities
across sites}
\usage{
networkModel(mods_list, n_bootstraps, n_cores)
}
\arguments{
\item{mods_list}{A \code{list} of models returned from \code{hurdleModel}}

\item{n_bootstraps}{\code{integer} representing the number of bootstrap replicates
to perform}

\item{n_cores}{Positive integer stating the number of processing cores to split the job across.
Default is \code{parallel::detect_cores() - 1}}
}
\description{
Run linear mixed models to identify predictors of species' eigencentralities
across sites
}
\details{
Species' network eigencentralities are used as the outcome variable in a linear mixed model
to identify environmental and biotic predictors. Only species with mean eigencentrality scores
in the top 25th percentile are considered, as these 'keystone' species are the ones that have the
strongest influences on community structure. The model pipeline is as follows:\cr
1: shuffle the dataset, impute any missing species' trait values and run a linear mixed model with
species and region as random effects (variable intercepts).
2: run a full model with all possible covariates as predictors
3: use stepwise variable selection (using AIC as the loss function) to identify key predictors
4: repeat \code{n_bootstraps} times to account for uncertainty in coefficients\cr
Models are then run to identify predictors of site-level interaction \emp{B}'os using a similar
pipeline
}
