% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBinomial.R
\name{plotBinomial}
\alias{plotBinomial}
\title{Construct regression plots of species' pairwise influences on occurrence probabilities
using model output from \code{lassoBinomial}}
\usage{
plotBinomial(lassoBinomial, cutoff)
}
\arguments{
\item{lassoBinomial}{A model object returned from \code{lassoBinomial_comm}}

\item{cutoff}{Positive numeric value indicating the minimum quantile of biotic coefficients'
absolute values (excluding \code{0}'s) that are considered as reflective of an
important influence on occurrence probability. Default is \code{0.05}, meaning that the top 95
percent of biotic coefficients are retained when performing the polynomial regression}
}
\value{
A \code{ggplot2} object visualising predicted habitat niche and phylogenetic distance
regression lines and 99% confidence intervals, along with the proportion of explained variance
captured by each predictor
}
\description{
Construct regression plots of species' pairwise influences on occurrence probabilities
using model output from \code{lassoBinomial}
}
\details{
Species' pairwise influences on occurrence probabilities are extracted from
\code{lassoBinomial$coefficients} and predicted
by their respective habitat niche and phylogenetic distances (extracted from
\code{data("Bird.ecol.distances")} and \code{data("Bird.phy.distances")}) using
a second-order polynomial regression of the form
\code{lm(Influence ~ poly(Phy.dist,2) + poly(Ecol.dist,2))}. Only pairwise
influences whose absolute magnitudes are above the specified \code{cutoff} quantile
are used in the regression
}
