% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictlassoBinomial.R
\name{predictlassoBinomial}
\alias{predictlassoBinomial}
\alias{lassoBinomial_metrics}
\title{Predict binomial outcomes from a fitted lassoBinomial_comm model}
\usage{
predictlassoBinomial(outcome_data, lassoBinomial, count_data, covariates,
  n_cores)

lassoBinomial_metrics(outcome_data, lassoBinomial, count_data, covariates,
  n_cores)
}
\arguments{
\item{outcome_data}{\code{dataframe}. The input data to be predicted, where the
columns represent binary occurrences of the species.
Colnames from this sample dataset must exactly match the colnames in the dataset that
was used to fit the \code{lassoBinomial_comm}}

\item{lassoBinomial}{A fitted \code{\link{lassoBinomial_comm}} model object}

\item{count_data}{A \code{dataframe} containing count observations
for species (each column representing a different species)}

\item{covariates}{A \code{matrix} of covariates, with \code{nrow(covariates) == nrow(response)}}

\item{n_cores}{Positive integer stating the number of processing cores to split the job across.
Default is \code{parallel::detect_cores() - 1}}
}
\value{
A \code{matrix} containing predictions for each observation in \code{outcome_data}.
}
\description{
This function calculates linear predictors for species' binary observations
using equations from a \code{\link{lassoBinomial_comm}} model.

\code{lassoBinomial_metrics} splits the data into ten folds and tests how well
the \code{lassoBinomial} model predicts binary observations in each fold.
\cr
\cr
Matrices of predictive metrics are returned.
}
