% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepDendrograms.R
\name{prepDendrograms}
\alias{prepDendrograms}
\title{Create a multivariate functional similarity dendrogram for species
using lists of trait datasets}
\usage{
prepDendrograms(func.datasets, prep.types)
}
\arguments{
\item{func.datasets}{A \code{list} containing dataframes of trait variables, where
rownames of each dataframe represent the species' names and columns represent trait variables.}

\item{prep.types}{A \code{list} describing which prepping method to apply to each dataset in
\code{func.datasets}. Each prep method must match one of the following:
\code{'prep.fuzzy'} (for proportional variables), \code{'prep.binary'} (for binary variables),
or \code{'Q'} (for continuous variables). For more information about prep methods, see
\code{\link[ade4]{dist.ktab}}}
}
\description{
Create a multivariate functional similarity dendrogram for species
using lists of trait datasets
}
\details{
Multiple functional trait datasets are incorporated with equal weights to
create a mixed-variable coefficient of distance, which is a generalized Gower's coefficient
of distance. This matrix is then scaled to range \code{[0, 1]} and returned.
}
\seealso{
\code{\link[ade4]{dist.ktab}}
}
