% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processNDVI.R
\name{processNDVI}
\alias{processNDVI}
\title{Extract mean summer NDVI values for buffers around lat/long points}
\usage{
processNDVI(filepath, points, buffer, n_cores, delete_files)
}
\arguments{
\item{filepath}{A \code{character} string stating the path to the folder where raw
Gimmsfiles are stored (downloaded using \code{\link[gimms]{downloadGimms}})}

\item{points}{A \code{dataframe} of points with columns \code{Latitude} and \code{Longitude}}

\item{buffer}{Positive numeric value stating the length (in meters) of the desired buffer radius.
Mean values of the climate variable will be calculated for each point from within this buffer}

\item{n_cores}{Positive integer stating the number of processing cores to split the job across.
Default is \code{parallel::detect_cores() - 1}}

\item{delete_files}{Logical describing whether or not to delete raw PRISM data folders once the job
is complete. Default is \code{FALSE}}
}
\description{
Extract mean summer NDVI values for buffers around lat/long points
}
