% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPRISM.R
\name{processPRISM}
\alias{processPRISM}
\title{Process mean monthly PRISM climate values into binned variables}
\usage{
processPRISM(monthly_climate_data)
}
\arguments{
\item{monthly_climate_data}{A \code{dataframe} of PRISM data for
points. Individual datasets should be created using \code{\link{meanPRISM}} and joined
together (e.g. with \code{\link[dplyr]{left_join}}), resulting in a single \code{dataframe}
with columns \code{Latitude}, \code{Longitude}, \code{Year}, \code{Month},
\code{Max.temp}, \code{Min.temp}, \code{Mean.temp} and \code{Precipitation}}
}
\description{
Process mean monthly PRISM climate values into binned variables
}
