% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USGS.bird.names.R
\docType{data}
\name{USGS.bird.names}
\alias{USGS.bird.names}
\title{USGS codes for avian species names}
\format{An object of class \code{data.frame} with 1175 rows and 9 columns.}
\usage{
data(USGS.bird.names)
}
\description{
A file containing species codes (used by BBS) and corresponding taxonomies
for birds in North America. This file is loaded in the \code{validateSpecies} script
}
\keyword{datasets}
