% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkBetaOS.R
\name{networkBetaOS}
\alias{networkBetaOS}
\title{Calculate turnover in predicted species interaction networks}
\usage{
networkBetaOS(adjacency_list, n_cores)
}
\arguments{
\item{adjacency_list}{\code{list}. The input data should be a list of
\code{\link[igraph]{graph.adjacency}} matrices, ideally as returned by
\code{\link[MRFcov]{predict_MRFnetworks}} using \code{metric == 'adjacency'}}

\item{n_cores}{Positive integer stating the number of processing cores to split the job across.
Default is \code{parallel::detect_cores() - 1}}
}
\value{
A \code{vector} containing predicted \emph{B}'os values for each network in \code{adjacency_list}
}
\details{
This function calculates \emph{B}'os,
i.e. the distance between a local network's realized interactions and the
potential interactions (calculated by integrating interactions from all local
networks into a regional \emph{metaweb}). Values close to \code{0} suggest there is a low local
selection of interactions (nearly all regional interactions are found in the local
network), whereas values closer to \code{1} reflect local selection of species' interactions
}
\references{
Poisot, T., Canard, E., Mouillot, D., Mouquet, N. & Gravel, D. (2012)
The dissimilarity of species interaction networks. Ecology Letters, 15, 1353-1361.
}
\seealso{
\code{\link[betalink]{beta_os_prime}}\cr
\code{\link[betalink]{metaweb}}
}
