% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBinomial.R
\name{plotBinomial}
\alias{plotBinomial}
\title{Construct regression plots of species' pairwise influences on occurrence probabilities
using model output from \code{lassoBinomial}}
\usage{
plotBinomial(lassoBinomial, cutoff)
}
\arguments{
\item{lassoBinomial}{A model object returned from \code{lassoBinomial}}

\item{cutoff}{Positive numeric value indicating the smallest magnitude a coefficient's
absolute value should be to be considered as an important influence on
occurrence probability. Default is \code{0.01}}
}
\value{
A \code{ggplot2} object visualising predicted habitat niche and phylogenetic distance
regression lines and 99% confidence intervals, along with the proportion of explained variance
captured by each predictor
}
\description{
Construct regression plots of species' pairwise influences on occurrence probabilities
using model output from \code{lassoBinomial}
}
\details{
Species' pairwise influences on occurrence probabilities are extracted from
\code{lassoBinomial$$Coefficients} and predicted
by their respective habitat niche and phylogenetic distances (extracted from
\code{data("Bird.ecol.distances")} and \code{data("Bird.phy.distances")}) using
a second-order polynomial regression of the form
\code{lm(Influence ~ poly(Phy.dist,2) + poly(Ecol.dist,2))}. Only pairwise
influences whose absolute magnitudes are above the specified \code{cutoff}
are used in the regression
}
