% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Storm.data.R
\docType{data}
\name{Storm.data}
\alias{Storm.data}
\title{Descriptions and locations of damaging storms in the USA}
\format{A \code{data.frame} containing 15 variables describing
location, severity and timing of the top third most damaging weather events recorded in the
NOAA storms database. Variables
include property damage, crop damage and total damage (all in US dollars), Year, Month,
Storm.type and geographrical coordinates}
\usage{
data(Storm.data)
}
\description{
A dataset on storm events was accessed from NOAA and processed using regex
to categorise storms into categories. The entire workflow is shown in the \code{Appendix 6} script.
}
\keyword{datasets}
