% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateSpecies.R
\name{validateSpecies}
\alias{validateSpecies}
\title{Determines if a species is "valid" by looking at its latin and}
\usage{
validateSpecies(species.df)
}
\description{
Determines if a species is "valid" by looking at its latin and
}
